backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
APScheduler==3.10.4
twilio==8.9.1
python-dotenv==1.0.0
